requirements.txt
requests>=2.31.0
PyYAML>=6.0.1
pandas>=2.2.2
bilibili-api-python>=16.1.0
SQLAlchemy>=2.0.32
